pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cache-clear"
DJANGO_SETTINGS_MODULE = "spy_cat_agency.settings"
testpaths = [
    "spycat",  # Add path to your app (this will include spycat/tests.py)
    "tests",  # Include other test directories if you have them
    "integration",  # Optional, if you have an integration folder
]



[tool.poetry]
name = "spy-cat-agency"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
django = "5.1.3"
djangorestframework = "3.15.2"
exceptiongroup = "1.2.2"
idna = "3.10"
iniconfig = "2.0.0"
packaging = "24.2"
pluggy = "1.5.0"
pytest = "8.3.3"
pytest-django = "4.9.0"  
requests = "2.32.3"
sqlparse = "0.5.2"
tomli = "2.1.0"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.7.2"
pre-commit = "^4.0.1"
mypy = "^1.13.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["."]
select = ["F", "I", "W", "E", "N", "D", "S", "ANN", "ARG", "B", "C4", "E", "ERA", "ISC", "PD", "PIE", "PL", "PT", "RET", "SIM", "TCH", "UP", "W", "BLE", "FBT", "B", "A", "COM", "DTZ", "ICN", "G", "INP", "T20", "RSE", "SLF", "PTH", "TD", "PGH", "PERF", "FURB", "RUF"]
ignore = []

[tool.mypy]
ignore_missing_imports = true
warn_redundant_casts = true
show_error_codes = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "spy_cat_agency.settings"
